requirements.txt
attrs==19.3.0
awsebcli==3.17.1
bcrypt==3.1.7
blessed==1.17.2
botocore==1.14.17
cached-property==1.5.1
cement==2.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.3.9
cryptography==2.8
docker==4.2.0
docker-compose==1.25.4
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
future==0.16.0
idna==2.7
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.9.5
jsonschema==3.2.0
MarkupSafe==1.1.1
marshmallow==3.5.1
paramiko==2.7.1
pathspec==0.5.9
pycparser==2.20
PyNaCl==1.3.0
pyrsistent==0.15.7
python-dateutil==2.8.0
python-dotenv==0.13.0
PyYAML==5.2
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
SQLAlchemy==1.3.16
termcolor==1.1.0
texttable==1.6.2
urllib3==1.24.3
virtualenv==16.7.9
wcwidth==0.1.8
websocket-client==0.57.0
Werkzeug==1.0.1
zipp==3.1.0
